# server/src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "UniVote API Documentation",
        "description": "REST API specification for the UniVote ACSES UMaT E-Voting System. Administrative actions require bearer tokens.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base Server"
        }
    ],
    "paths": {
...
        "/elections": {
            "get": {
                "summary": "List all elections",
                "responses": {
                    "200": {
                        "description": "Array of configured elections.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Election"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new election",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "ACSES Presidentials 2026"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Presidential elections for CSE department."
                                    },
                                    "departmentId": {
                                        "type": "string",
                                        "example": "dept-cse"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Student Representative"
                                    },
                                    "startTime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "endTime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "eligibleVoterCount": {
                                        "type": "number",
                                        "example": 350
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "President",
                                            "Vice President",
                                            "Secretary"
                                        ]
                                    }
                                },
                                "required": [
                                    "title",
                                    "startTime",
                                    "endTime"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Election created as Draft status."
                    }
                }
            }
        },
        "/elections/{id}": {
            "patch": {
                "summary": "Update election specifications",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "active",
                                            "closed"
                                        ]
                                    },
                                    "startTime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "endTime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "eligibleVoterCount": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Election settings updated successfully."
                    },
                    "404": {
                        "description": "Election not found."
                    }
                }
            }
        },
        "/elections/{id}/categories": {
            "post": {
                "summary": "Add target ballot categories",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Financial Secretary"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ballot category created."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "Election": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "startTime": {
                        "type": "string"
                    },
                    "endTime": {
                        "type": "string"
                    }
                }
            }
        }
    }
}